+++ src/itinerary/index.tsx
@@ -53,7 +53,7 @@ const StyledItinerary = styled(Itinerary)`
   }
 
   &
-    .kirk-itinerary--small
+    .kirk-itinerary--noTime
     .kirk-itinerary-location:not(.kirk-itinerary--arrival)
     .kirk-itinerary-location-city {
     padding-bottom: 0;
@@ -75,7 +75,7 @@ const StyledItinerary = styled(Itinerary)`
     left: calc(${gutterPaddingStart} + ${componentSizes.timeWidth});
   }
 
-  & .kirk-itinerary--small .kirk-itinerary-location .kirk-itinerary-location-city::before {
+  & .kirk-itinerary--noTime .kirk-itinerary-location .kirk-itinerary-location-city::before {
     left: 2px;
   }
 
@@ -94,7 +94,7 @@ const StyledItinerary = styled(Itinerary)`
     top: calc(4px + ${space.m} + ${gutterTopOffset});
   }
 
-  & .kirk-itinerary--small .kirk-itinerary-location .kirk-itinerary-location-city::after {
+  & .kirk-itinerary--noTime .kirk-itinerary-location .kirk-itinerary-location-city::after {
     left: 0;
   }
 
@@ -129,8 +129,8 @@ const StyledItinerary = styled(Itinerary)`
     bottom: -4px;
   }
 
-  & .kirk-itinerary--small .kirk-itinerary-fromAddon::before,
-  & .kirk-itinerary--small .kirk-itinerary-location.kirk-itinerary-location--toAddon::before {
+  & .kirk-itinerary--noTime .kirk-itinerary-fromAddon::before,
+  & .kirk-itinerary--noTime .kirk-itinerary-location.kirk-itinerary-location--toAddon::before {
     left: 2px;
   }
 
@@ -154,8 +154,8 @@ const StyledItinerary = styled(Itinerary)`
     left: calc(${gutterPaddingStart} + ${componentSizes.timeWidth} - 2px);
   }
 
-  & .kirk-itinerary--small .kirk-itinerary-fromAddon::after,
-  & .kirk-itinerary--small .kirk-itinerary-location.kirk-itinerary-location--toAddon::after {
+  & .kirk-itinerary--noTime .kirk-itinerary-fromAddon::after,
+  & .kirk-itinerary--noTime .kirk-itinerary-location.kirk-itinerary-location--toAddon::after {
     left: 0;
   }
 
@@ -176,8 +176,8 @@ const StyledItinerary = styled(Itinerary)`
     left: calc(${gutterPaddingStart} + ${componentSizes.timeWidth});
   }
 
-  & .kirk-itinerary--small .kirk-itinerary-fromAddon .kirk-itinerary-addon-content,
-  & .kirk-itinerary--small .kirk-itinerary-toAddon .kirk-itinerary-addon-content {
+  & .kirk-itinerary--noTime .kirk-itinerary-fromAddon .kirk-itinerary-addon-content,
+  & .kirk-itinerary--noTime .kirk-itinerary-toAddon .kirk-itinerary-addon-content {
     left: 0;
   }