+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[FIX]** Fix `Itinerary` component display without time
 - [...]
 
 # V17.0.0 (24/12/2019)